pyproject.toml
[tool.poetry]
name = "fastapi-blog"
version = "0.1.0"
description = "Mini blog on FastAPI"
authors = ["Vanya <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {version = "0.115.0", extras = ["all"]}
pydantic = "2.9.2"
pydantic-settings = "2.5.2"
uvicorn = "0.31.0"
jinja2 = "3.1.4"
aiosqlite = "0.20.0"
alembic = "1.13.3"
sqlalchemy = "2.0.35"
bcrypt = "4.0.1"
passlib = {version = "1.7.4", extras = ["bcrypt"]}
python-jose = "3.3.0"
loguru = "0.7.2"
markdown2 = "2.5.1"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.3"
pytest-asyncio = "0.24.0"
pytest-cov = "^7.0.0"
pre-commit = "3.7.0"
mypy = "^1.13.0"
black = "^24.1.0"
isort = "^5.12.0"
ruff = "^0.12.0"

[tool.mypy]
python_version = "3.13"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
no_implicit_optional = true
check_untyped_defs = false
disallow_any_generics = false
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]


